+++ modules/farm_rothamsted_quick/src/Traits/QuickQuantityFieldTrait.php
@@ -57,7 +57,7 @@ trait QuickQuantityFieldTrait {
         '#type' => 'number',
         '#weight' => 5,
         '#min' => 0,
-        '#step' => 0.01,
+        '#step' => 0.0001,
       ],
       'units' => [
         '#weight' => 10,