+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow 4 decimal points in quick form quantity values. [#463](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/463)
 - Move and rename proposal requested field location. [#446](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/446)
 - Rename drilling year to planting year in experiment plan. [#457](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/457)
+- Rename grant code to funder grant code. [#462](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/462)
 
 ## 2.11.3 2023-05-30